refactor(podman): extract MacOSInstaller to a dedicated file
